Bruh you ever get that feeling where you thought you were on top of everything and life was good, and then you realised that your planning was insufficient and you glossed over how darts is actually played and now you need to rework the core loop of the game. Shouldn't be too hard, I just need to rename some variables to make more sense and then encapsulate the current game loop in some other loops which represents sets and matches. Sighhhhhhh...

+++ Unit 2 - Darts/Player.cpp
@@ -182,17 +182,17 @@ void Player::ResetBust()
 	BustFlag = false; 
 }
 
-uint16_t Player::GetLifetimethrows()
+uint32_t Player::GetLifetimethrows()
 {
 	return lifetimeThrows;
 }
 
-uint16_t Player::GetLifetimeBulls()
+uint32_t Player::GetLifetimeBulls()
 {
 	return lifetimeBulls;
 }
 
-uint16_t Player::GetLifetimeprecisehits()
+uint32_t Player::GetLifetimeprecisehits()
 {
 	return lifetimePreciseHits;
 }
@@ -202,7 +202,7 @@ void Player::IncrementLifetimeWins()
 	lifetimeTotalWins++;
 }
 
-uint16_t Player::GetLifetimeWins()
+uint32_t Player::GetLifetimeWins()
 {
 	return lifetimeTotalWins;
 }
